pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "finkernel"
version = "0.1.0"
description = "Policy-first financial agent middleware MVP"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.135,<0.136",
  "uvicorn>=0.44,<0.45",
  "sqlalchemy>=2.0,<2.1",
  "pydantic-settings>=2.13,<3",
  "httpx>=0.28,<0.29",
  "psycopg[binary]>=3.3,<3.4",
  "redis>=7.4,<8",
  "discord.py>=2.7,<3",
  "mcp>=1,<2",
  "pgvector>=0.4,<0.5",
]

[project.optional-dependencies]
dev = [
  "pytest>=9,<10",
  "pytest-asyncio>=1.3,<2",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
addopts = "-q"
